lab-1-socket-conn-custom-threadpool-grpc/src/requirements.txt
asgiref==3.6.0
grpcio==1.51.1
grpcio-tools==1.51.1
numpy==1.24.1
pandas==1.5.3
protobuf==3.20.1
Pyro5==5.14
python-dateutil==2.8.2
pytz==2022.7.1
serpent==1.41
six==1.16.0
sqlparse==0.4.3
